requirements.txt
# HubSpot API
httpx>=0.24.0
python-dotenv>=1.0.0

# データベース
aiomysql>=0.2.0
pymysql>=1.1.0
cryptography>=41.0.0

# Ollama API
ollama>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.2.1

# Authentication
authlib>=1.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Session Management
itsdangerous>=2.1.2

# ベクトルDB
chromadb>=0.4.22

# その他
asyncio>=3.4.3
logging>=0.4.9.6

